+++ .golangci.yml
@@ -28,6 +28,6 @@ linters:
 
 linters-settings:
   goimports:
-    local-prefixes: github.com/karamaru-alpha/hoodie
+    local-prefixes: github.com/karamaru-alpha/days
   exhaustive:
     ignore-enum-members: "UNKNOWN"
