+++ config/api/local.go
@@ -1,6 +1,6 @@
 package api
 
-import "github.com/karamaru-alpha/hoodie/pkg/domain/config"
+import "github.com/karamaru-alpha/days/pkg/domain/config"
 
 var localAPIConfig = &config.APIConfig{
 	Env:  config.EnvLocal,
